Changed some size_t's back to ints. This mostly concerns those that refer to geometrical dimensions and on which substraction operations can reasonably be made.

+++ lib/assembly/discrete_inverse_sparse_linear_operator.cpp
@@ -56,9 +56,9 @@ void solveWithAmesos<double>(Epetra_LinearProblem& problem,
                              arma::Mat<double>& armaSolution,
                              const arma::Mat<double>& armaRhs)
 {
-    const int rowCount = armaRhs.n_rows;
+    const size_t rowCount = armaRhs.n_rows;
     assert(rowCount == armaSolution.n_rows);
-    const int rhsCount = armaRhs.n_cols;
+    const size_t rhsCount = armaRhs.n_cols;
     assert(rhsCount == armaSolution.n_cols);
 
     Epetra_Map map(rowCount, 0 /* base index */, Epetra_SerialComm());
