+++ lib/fiber/kernel.hpp
@@ -41,9 +41,9 @@ public:
 
     virtual ~Kernel() {}
 
-    virtual size_t worldDimension() const = 0;
-    virtual size_t domainDimension() const = 0;
-    virtual size_t codomainDimension() const = 0;
+    virtual int worldDimension() const = 0;
+    virtual int domainDimension() const = 0;
+    virtual int codomainDimension() const = 0;
 
     virtual void addGeometricalDependencies(size_t& testGeomDeps,
                                             size_t& trialGeomDeps) const = 0;
