+++ lib/fiber/laplace_3d_adjoint_double_layer_potential_kernel.cpp
@@ -70,14 +70,14 @@ void Laplace3dAdjointDoubleLayerPotentialKernel<ValueType>::evaluateAtPointPairs
     const arma::Mat<CoordinateType>& testNormals = testGeomData.normals;
 
 #ifndef NDEBUG
-    const size_t worldDim = worldDimension();
-    if (testPoints.n_rows != worldDim || trialPoints.n_rows != worldDim)
+    const int worldDim = worldDimension();
+    if ((int)testPoints.n_rows != worldDim || (int)trialPoints.n_rows != worldDim)
         throw std::invalid_argument("Laplace3dAdjointDoubleLayerPotentialKernel::evaluateAtPointPairs(): "
                                     "3D coordinates required");
     if (testPoints.n_cols != trialPoints.n_cols)
         throw std::invalid_argument("Laplace3dAdjointDoubleLayerPotentialKernel::evaluateAtPointPairs(): "
                                     "number of test and trial points must be equal");
-    assert(testNormals.n_rows == worldDim);
+    assert((int)testNormals.n_rows == worldDim);
     assert(testNormals.n_cols == testPoints.n_cols);
 #endif
 
@@ -100,11 +100,11 @@ void Laplace3dAdjointDoubleLayerPotentialKernel<ValueType>::evaluateOnGrid(
     const arma::Mat<CoordinateType>& testNormals = testGeomData.normals;
 
 #ifndef NDEBUG
-    const size_t worldDim = worldDimension();
-    if (testPoints.n_rows != worldDim || trialPoints.n_rows != worldDim)
+    const int worldDim = worldDimension();
+    if ((int)testPoints.n_rows != worldDim || (int)trialPoints.n_rows != worldDim)
         throw std::invalid_argument("Laplace3dAdjointDoubleLayerPotentialKernel::evaluate(): "
                                     "3D coordinates required");
-    assert(testNormals.n_rows == worldDim);
+    assert((int)testNormals.n_rows == worldDim);
     assert(testNormals.n_cols == testPoints.n_cols);
 #endif
