+++ lib/fiber/laplace_3d_adjoint_double_layer_potential_kernel.hpp
@@ -46,9 +46,9 @@ class Laplace3dAdjointDoubleLayerPotentialKernel : public Kernel<ValueType>
 public:
     typedef typename Kernel<ValueType>::CoordinateType CoordinateType;
 
-    virtual size_t worldDimension() const { return 3; }
-    virtual size_t domainDimension() const { return 1; }
-    virtual size_t codomainDimension() const { return 1; }
+    virtual int worldDimension() const { return 3; }
+    virtual int domainDimension() const { return 1; }
+    virtual int codomainDimension() const { return 1; }
 
     virtual void addGeometricalDependencies(size_t& testGeomDeps,
                                             size_t& trialGeomDeps) const;
