+++ lib/fiber/modified_helmholtz_3d_single_layer_potential_kernel.cpp
@@ -73,8 +73,8 @@ void ModifiedHelmholtz3dSingleLayerPotentialKernel<ValueType>::evaluateAtPointPa
     const arma::Mat<CoordinateType>& trialPoints = trialGeomData.globals;
 
 #ifndef NDEBUG
-    if (testPoints.n_rows != worldDimension() ||
-            trialPoints.n_rows != worldDimension())
+    if ((int)testPoints.n_rows != worldDimension() ||
+            (int)trialPoints.n_rows != worldDimension())
         throw std::invalid_argument("Laplace3dSingleLayerPotentialKernel::evaluateAtPointPairs(): "
                                     "3D coordinates required");
     if (testPoints.n_cols != trialPoints.n_cols)
@@ -99,8 +99,8 @@ void ModifiedHelmholtz3dSingleLayerPotentialKernel<ValueType>::evaluateOnGrid(
     const arma::Mat<CoordinateType>& trialPoints = trialGeomData.globals;
 
 #ifndef NDEBUG
-    if (testPoints.n_rows != worldDimension() ||
-            trialPoints.n_rows != worldDimension())
+    if ((int)testPoints.n_rows != worldDimension() ||
+            (int)trialPoints.n_rows != worldDimension())
         throw std::invalid_argument("Laplace3dSingleLayerPotentialKernel::evaluate(): "
                                     "3D coordinates required");
 #endif
