+++ lib/fiber/numerical_test_trial_integrator_imp.hpp
@@ -77,9 +77,9 @@ void NumericalTestTrialIntegrator<BasisFunctionType, ResultType, GeometryFactory
     // elementCount != 0, set elements of result to 0.
 
     // Evaluate constants
-    const size_t componentCount = m_testExpression.codomainDimension();
-    const size_t testDofCount = testBasis.size();
-    const size_t trialDofCount = trialBasis.size();
+    const int componentCount = m_testExpression.codomainDimension();
+    const int testDofCount = testBasis.size();
+    const int trialDofCount = trialBasis.size();
 
     if (m_trialExpression.codomainDimension() != componentCount)
         throw std::runtime_error("NumericalTestTrialIntegrator::integrate(): "
@@ -113,12 +113,12 @@ void NumericalTestTrialIntegrator<BasisFunctionType, ResultType, GeometryFactory
         m_testExpression.evaluate(testBasisData, geomData, testValues);
         m_trialExpression.evaluate(trialBasisData, geomData, trialValues);
 
-        for (size_t trialDof = 0; trialDof < trialDofCount; ++trialDof)
-            for (size_t testDof = 0; testDof < testDofCount; ++testDof)
+        for (int trialDof = 0; trialDof < trialDofCount; ++trialDof)
+            for (int testDof = 0; testDof < testDofCount; ++testDof)
             {
                 ResultType sum = 0.;
                 for (size_t point = 0; point < pointCount; ++point)
-                    for (size_t dim = 0; dim < componentCount; ++dim)
+                    for (int dim = 0; dim < componentCount; ++dim)
                         sum +=  m_quadWeights[point] *
                                 geomData.integrationElements(point) *
                                 conjugate(testValues(dim, testDof, point)) *
